Migrate to @base-ui/react 1.6 (drop @base-ui-components RC) + dedupe challenge primitives
Base UI migration:
- Rewrite all 46 imports of @base-ui-components/react@1.0.0-rc.0 (popover,
  dialog, menu, select, context-menu, alert-dialog, switch, combobox, tabs,
  checkbox, accordion) to @base-ui/react and remove the RC package, ending
  the two-copies-of-the-same-library situation.
- Bump @base-ui/react ^1.4.1 -> ^1.6.0 (latest stable).
- Audited every breaking change in the rc.0->1.6.0 changelog window: all are
  in components this app doesn't use (OTP Field) or already absorbed (package
  rename, Drawer un-preview <=1.4.1). The data-attribute contracts the CSS
  relies on (data-highlighted, data-selected, data-popup-open, ...) are
  unchanged for the components in use; the Tabs attr renames happened in
  1.0.0-beta.5, before the RC this app was on.

Challenge-family dedup (ChallengeCard, SqlChallengeCard, CodeBlock,
SqlCodeBlock): byte-identical private copies of Status/TestState/
DisplayedTest, detectIsMac, MIN_RUN_OVERLAY_MS, lineNumbersWithOffset,
LanguageGlyph, and useBlockId (verified by hash) now live once in
challengeShared.tsx.

Verified: tsc --noEmit, eslint, all 949 unit tests; Base UI-heavy e2e specs
(mobile menus, SQL create-object dialogs across SQLite/PostgreSQL/DuckDB)
run in a real browser.
