+++ app/_components/ErDiagramPane.tsx
@@ -23,9 +23,9 @@ import type { TableColumnInfo, ForeignKeyInfo } from "./runtime/sqlite";
 import { MdOutlineKey } from "react-icons/md";
 import { IoLink } from "react-icons/io5";
 import { ChevronRight, Table } from "lucide-react";
-import { ContextMenu } from "@base-ui-components/react/context-menu";
-import { Menu } from "@base-ui-components/react/menu";
-import { Popover } from "@base-ui-components/react/popover";
+import { ContextMenu } from "@base-ui/react/context-menu";
+import { Menu } from "@base-ui/react/menu";
+import { Popover } from "@base-ui/react/popover";
 
 // ────────────────────────────────────────────────────────────────────────────
 // Contexts, table action callbacks + node selection state
