+++ app/_components/ai/AskAiWidget.tsx
@@ -18,7 +18,7 @@ import {
 } from "react";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
-import { Popover } from "@base-ui-components/react/popover";
+import { Popover } from "@base-ui/react/popover";
 import {
   Sparkles,
   X,
