+++ app/_components/cloud/ShareControls.tsx
@@ -19,7 +19,7 @@
  */
 
 import { useCallback, useEffect, useState } from "react";
-import { Dialog } from "@base-ui-components/react/dialog";
+import { Dialog } from "@base-ui/react/dialog";
 import { Check, Copy as CopyIcon, Loader2, Share2 } from "lucide-react";
 
 import { useSession } from "@/lib/auth/client";
