+++ app/_components/sql/components/DdlViewerDialog.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Dialog } from "@base-ui-components/react/dialog";
+import { Dialog } from "@base-ui/react/dialog";
 import type { ReactNode } from "react";
 import { DdlViewer } from "./DdlViewer";
