+++ app/_components/sql/components/RenameDatabaseDialog.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Dialog } from "@base-ui-components/react/dialog";
+import { Dialog } from "@base-ui/react/dialog";
 
 export interface ExtensionOption {
   value: string;
