+++ app/_components/sql/components/SchemaActionDialogs.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import type { ReactNode } from "react";
-import { AlertDialog } from "@base-ui-components/react/alert-dialog";
+import { AlertDialog } from "@base-ui/react/alert-dialog";
 
 export interface SchemaActionDialogsProps {
   dropEntityPending: { kind: string; name: string } | null;
