+++ app/_components/sql/components/SchemaItem.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import { memo, useState, useCallback, useMemo, useRef, useEffect } from "react";
-import { Popover } from "@base-ui-components/react/popover";
-import { ContextMenu } from "@base-ui-components/react/context-menu";
-import { Menu } from "@base-ui-components/react/menu";
+import { Popover } from "@base-ui/react/popover";
+import { ContextMenu } from "@base-ui/react/context-menu";
+import { Menu } from "@base-ui/react/menu";
 import { ChevronDown, ChevronRight, Plus, Search, Table, View } from "lucide-react";
 import { IoLink } from "react-icons/io5";
 import { MdOutlineKey } from "react-icons/md";
