+++ app/_components/sql/components/StructureCombobox.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState, useRef, useEffect, useMemo } from "react";
-import { Combobox } from "@base-ui-components/react/combobox";
+import { Combobox } from "@base-ui/react/combobox";
 import { ChevronDown } from "lucide-react";
 import {
   computeVisibleTypeGroups,
