+++ package.json
@@ -29,8 +29,7 @@
     "test:solutions": "playwright test e2e/challenge-solutions.spec.ts"
   },
   "dependencies": {
-    "@base-ui-components/react": "^1.0.0-rc.0",
-    "@base-ui/react": "^1.4.1",
+    "@base-ui/react": "^1.6.0",
     "@codemirror/autocomplete": "^6.20.1",
     "@codemirror/commands": "^6.10.3",
     "@codemirror/lang-cpp": "^6.0.3",
